sanza/ci/drone_project/requirements.txt
psycopg2
django-localeurl
django-modeltranslation
